# src/web/openapi.json
{"openapi":"3.1.0","info":{"title":"Sidefeed API","version":"0.1.0","description":"Aggregate, filter, search, exchange, and publish feeds."},"servers":[{"url":"/"}],"components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer"}},"schemas":{"Error":{"type":"object","required":["error"],"properties":{"error":{"type":"string"}}}}},"paths":{"/healthz":{"get":{"summary":"Liveness","responses":{"200":{"description":"Healthy"}}}},"/readyz":{"get":{"summary":"Database readiness","responses":{"200":{"description":"Ready"}}}},"/api/v1/sources":{"get":{"summary":"List sources","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Sources"}}},"post":{"summary":"Create source","security":[{"bearerAuth":[]}],"responses":{"201":{"description":"Created"}}}},"/api/v1/sources/{id}/poll":{"post":{"summary":"Poll source now","security":[{"bearerAuth":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Import result"}}}},"/api/v1/import/opml":{"post":{"summary":"Import OPML","security":[{"bearerAuth":[]}],"responses":{"201":{"description":"Sources created"}}}},"/api/v1/feeds":{"get":{"summary":"List feeds","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Feeds"}}},"post":{"summary":"Create feed","security":[{"bearerAuth":[]}],"responses":{"201":{"description":"Created"}}}},"/api/v1/feeds/{slug}/items":{"get":{"summary":"Page through feed items","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"}},{"name":"limit","in":"query","schema":{"type":"integer","minimum":1,"maximum":200}},{"name":"cursor","in":"query","schema":{"type":"string"}},{"name":"tag","in":"query","schema":{"type":"string"},"description":"Comma separated tags"},{"name":"matching","in":"query","schema":{"type":"string","enum":["any","all"]}}],"responses":{"200":{"description":"Item page"}}}},"/api/v1/feeds/{slug}/search":{"get":{"summary":"Full-text search","responses":{"200":{"description":"Results"}}}},"/api/v1/feeds/{slug}/semantic":{"get":{"summary":"Hybrid vector and full-text search","responses":{"200":{"description":"Ranked results"}}}},"/api/v1/feeds/{slug}/stream":{"get":{"summary":"Server-sent item stream","responses":{"200":{"description":"SSE stream"}}}},"/api/v1/peers":{"get":{"summary":"List peers","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Peers"}}},"post":{"summary":"Create peer","security":[{"bearerAuth":[]}],"responses":{"201":{"description":"Created"}}}},"/feeds/{slug}.rss":{"get":{"summary":"RSS output","responses":{"200":{"description":"RSS document"}}}},"/feeds/{slug}.json":{"get":{"summary":"JSON Feed output","responses":{"200":{"description":"JSON Feed document"}}}},"/feeds/{slug}/newsletter":{"get":{"summary":"Newsletter HTML or text","responses":{"200":{"description":"Rendered digest"}}}},"/feeds/{slug}/thread.json":{"get":{"summary":"Social thread payload","responses":{"200":{"description":"Thread posts"}}}},"/api/v1/public/feeds":{"get":{"summary":"List public feeds","description":"Read-only feed index used by the unauthenticated reader interface. Never lists private feeds.","responses":{"200":{"description":"Public feeds"}}}},"/api/v1/recent":{"get":{"summary":"Recent items across every public feed","description":"Ranked by recency and source variety with a per-feed cap, so one busy source cannot fill it.","parameters":[{"name":"hours","in":"query","schema":{"type":"integer","minimum":1,"maximum":336}},{"name":"limit","in":"query","schema":{"type":"integer","minimum":1,"maximum":200}},{"name":"tag","in":"query","schema":{"type":"string"},"description":"Comma separated tags"},{"name":"matching","in":"query","schema":{"type":"string","enum":["any","all"]}}],"responses":{"200":{"description":"Ranked items"}}}},"/api/v1/search":{"get":{"summary":"Full-text search across public feeds","description":"FTS5 with each word quoted and prefix matched, so punctuation cannot break the query. Without hours the whole archive is searched.","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string"}},{"name":"feed","in":"query","schema":{"type":"string"},"description":"Restrict to one feed slug"},{"name":"hours","in":"query","schema":{"type":"integer"}},{"name":"order","in":"query","schema":{"type":"string","enum":["newest","oldest"]}},{"name":"limit","in":"query","schema":{"type":"integer","minimum":1,"maximum":200}},{"name":"tag","in":"query","schema":{"type":"string"},"description":"Comma separated tags"},{"name":"matching","in":"query","schema":{"type":"string","enum":["any","all"]}}],"responses":{"200":{"description":"Matches, most relevant first"}}}},"/api/v1/tags":{"get":{"summary":"Tag counts across public feeds","parameters":[{"name":"hours","in":"query","schema":{"type":"integer"}},{"name":"limit","in":"query","schema":{"type":"integer"}}],"responses":{"200":{"description":"Tags with counts"}}}},"/api/v1/feeds/{slug}/tags":{"get":{"summary":"Tag counts for one feed","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"}},{"name":"limit","in":"query","schema":{"type":"integer"}}],"responses":{"200":{"description":"Tags with counts"}}}},"/api/v1/items/{id}/enrich":{"post":{"summary":"Derive tags and a summary now","security":[{"bearerAuth":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"201":{"description":"Tags and summary"}}}}}}